Moved all typelookupservice files to src
