+++ src/cpp/fastdds/builtin/typelookupservice/TypeLookupManager.cpp
@@ -17,7 +17,7 @@
  *
  */
 
-#include <fastdds/dds/builtin/typelookup/TypeLookupManager.hpp>
+#include <fastdds/builtin/typelookupservice/TypeLookupManager.hpp>
 
 #include <fastdds/rtps/builtin/BuiltinProtocols.h>
 #include <fastdds/rtps/builtin/data/ParticipantProxyData.h>
