+++ src/cpp/fastdds/builtin/typelookupservice/TypeLookupReplyListener.cpp
@@ -17,8 +17,8 @@
  *
  */
 
-#include <fastdds/dds/builtin/typelookup/TypeLookupReplyListener.hpp>
-#include <fastdds/dds/builtin/typelookup/TypeLookupManager.hpp>
+#include <fastdds/builtin/typelookupservice/TypeLookupReplyListener.hpp>
+#include <fastdds/builtin/typelookupservice/TypeLookupManager.hpp>
 #include <fastdds/rtps/participant/RTPSParticipantListener.h>
 #include <fastdds/rtps/builtin/BuiltinProtocols.h>
