+++ src/cpp/fastdds/builtin/typelookupservice/common/RPCHeadersImpl.cpp
@@ -17,8 +17,8 @@
  *
  */
 
-#include <fastdds/dds/builtin/common/ReplyHeader.hpp>
-#include <fastdds/dds/builtin/common/RequestHeader.hpp>
+#include <fastdds/builtin/typelookupservice/common/ReplyHeader.hpp>
+#include <fastdds/builtin/typelookupservice/common/RequestHeader.hpp>
 #include <fastdds/rtps/common/CdrSerialization.hpp>
 
 #include "RPCHeadersImplCdrAux.ipp"
