+++ src/cpp/fastdds/builtin/typelookupservice/common/TypeLookupTypesCdrAux.ipp
