+++ src/cpp/rtps/builtin/discovery/participant/PDPSimple.cpp
@@ -20,7 +20,7 @@
 
 #include <mutex>
 
-#include <fastdds/dds/builtin/typelookup/TypeLookupManager.hpp>
+#include <fastdds/builtin/typelookupservice/TypeLookupManager.hpp>
 #include <fastdds/dds/log/Log.hpp>
 #include <fastdds/rtps/builtin/BuiltinProtocols.h>
 #include <fastdds/rtps/builtin/data/NetworkConfiguration.hpp>
