+++ test/mock/rtps/TypeLookupManager/fastdds/dds/builtin/typelookup/TypeLookupManager.hpp
@@ -19,7 +19,7 @@
 #include <vector>
 #include <gmock/gmock.h>
 
-#include <fastdds/dds/builtin/typelookup/common/TypeLookupTypes.hpp>
+#include <fastdds/builtin/typelookupservice/common/TypeLookupTypes.hpp>
 
 namespace eprosima {
